+++ tests/testthat/test-utils-xportr.R
@@ -140,3 +140,19 @@ test_that("utils-xportr Test 13: Get error message when the length of a characte
     "Length of A must be 200 bytes or less."
   )
 })
+
+## Test 14: xpt_validate: Get error message when the class of a datetime variable is numeric ----
+test_that("utils-xportr Test 14: Get error message when the class of a datetime variable is numeric", { # nolint
+  adlb <- data.frame(
+    USUBJID = c("P01", "P02"),
+    TRTSDT = as.Date(c("2023-01-01", "2023-02-15")),
+    TRTSTM = as.POSIXct(c("2023-01-01 10:00:00", "2023-02-15 11:30:00")),
+    TRTEDTM = as.numeric(as.POSIXct(c("2023-01-01 10:00:00", "2023-02-15 11:30:00"))),
+    LBDTC = c("2023-01-01", "2023-02-15"),
+    AVAL = c(10, 20)
+  )
+  expect_equal(
+    xpt_validate(adlb),
+    "Variable `TRTEDTM` do not have an R 'date', 'datetime' or 'time' class:date, POSIXct/lt, hms or difftime."
+  )
+})